requirements.txt
python-telegram-bot==13.15
schedule==1.1.0
requests==2.28.2
python-dotenv==1.0.0
cachetools==4.2.2 